transaction_tracker: add more information to the out-of-memory message
A Slack user reported seeing an error like this:

  transaction_tracker.cc:119] Transaction failed, tablet \
    00000000000000000000000000000000 transaction memory consumption (0) has \
    exceeded its limit (67108864) or the limit of an ancestral tracker

This doesn't make sense; how could the limit be exceeded if the consumption
is 0? As far as I can tell, it can only happen if the failed transaction had
a colossal footprint of its own.

This patch changes the generated error to include the footprint of the
transaction itself.

Change-Id: Ic4017c1aecec29cd4b4b04c15efa4bf38c925cb9
Reviewed-on: http://gerrit.cloudera.org:8080/11765
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>