KUDU-2584: Prevent flaky off-by-one errors in backup tests
This patch adds 1 ms to the target snapshot time when
a backup is taken. This ensures that we don’t have
flakes due to off-by-one errors where all the values are not read.

The underlying reason for adding 1 ms is that we pass
the timestamp in ms granularity but the snapshot time
consists of microseconds plus a logical clock. This
means if the data is inserted with a fraction of a ms
remaining it could be truncated and unread.

Additionaly this patch copies over the timestamp
propagation call from the KuduRDD and ensures
the Spark tests use the Kudu client from the
KuduContext. This should further prevent future
snapshot issues.

This patch also includes an auto-formating change in
KuduBackupOptions that must have been missed in
a previous commit.

Change-Id: Ia0f1b4a4138cc8c913543a68fad748927cdc439d
Reviewed-on: http://gerrit.cloudera.org:8080/11815
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>