thirdparty: work around an upstream LLVM build race
There is a long-standing race[1] in LLVM's build that seems to only manifest
when the compiler-rt LLVM project (used for sanitizers and the like) is
built via make with certain values of -j.

I spent some quality time with the compiler-rt build but was unable to find
its root cause, so instead let's work around it in our own build.

1. https://bugs.llvm.org/show_bug.cgi?id=26054

Change-Id: Ia3bf2a171b85d34326e187a4d14dbffe7326100d
Reviewed-on: http://gerrit.cloudera.org:8080/11834
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins