[cmake] place all DATA_FILES under bin/testdata
Updated the convention for the DATA_FILES parameter of the ADD_KUDU_TEST
macro. With this patch, all test data files are placed under the
<build_dir>/bin/testdata directory after running cmake.

The new convention eases the restrictions on the layout of the source
data files. That, in its turn, helps to stop duplicating auxiliary files
needed for various tests.

Change-Id: I75064c6bcf9a569648b922b55f539730769cdf4c
Reviewed-on: http://gerrit.cloudera.org:8080/11841
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>