[test] optional table creation in TabletServerIntegrationTestBase
Creation of the test table in TabletServerIntegrationTestBase is now
optional.  If not specified otherwise, the test table is created by
default, preserving the original behavior.

The motivation for this change is that sometimes a derived test
might need the functionality of the TabletServerIntegrationTestBase
but expects not a single table created after SetUp().

The updated behavior will be used in a follow-up patch.

Change-Id: I818a2eaf76fdee11ae792c27eddae311b775efeb
Reviewed-on: http://gerrit.cloudera.org:8080/11861
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins