Disable rowset compaction in TestUndoDeltaBlockGc
In a follow-up to this patch, rowset compaction is enhanced to merge
small rowsets. This will cause TestUndoDeltaBlockGc to be flaky, since
it creates small rowsets with undo deltas and tries to wait until they
are GC'd by the UndoDeltaBlockGc maintenance op; however, if rowset
compaction merges the rowsets after the undos have passed the AHM but
before UndoDeltaBlockGc runs, no undos will be deleted by the
UndoDeltaBlockGc task and the test will fail.

This fixes the test by simply disabling rowset compactions.

I ran the test 100 times with the change and the KUDU-1400 change on
top and saw zero failures.

Change-Id: If23f4568d10d83b7f463663cfc0d4052f2602224
Reviewed-on: http://gerrit.cloudera.org:8080/11884
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>