Limit number of rowsets in compaction selection to 32 in TSAN mode
TSAN limits the number of simultaneous lock acquisitions in a single
thread to 64 when using the deadlock detector[1]. However, compaction
can select up to 128 (128MB budget / 1MB min rowset size) rowsets in a
single op. kudu-tool-test's TestNonRandomWorkloadLoadgen almost always
hits TSAN's limit when the KUDU-1400 changes following this patch are
applied. This patch prevents this by limiting the number of rowsets
selected for a compaction to 32 when running under TSAN.

I ran the test with the KUDU-1400 changes on top and saw 97/100
failures. With the change, I saw 100 successes.

[1]: https://github.com/google/sanitizers/issues/950

Change-Id: I01ad4ba3a13995c194c3308d72c1eb9b611ef766
Reviewed-on: http://gerrit.cloudera.org:8080/11885
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>