docs: Use a non-SSH method of installing Gerrit commit hook
Apparently, contributors in China cannot connect to the Gerrit SSH port.
Because of this, they cannot follow the How to Contribute directions for
installing the Gerrit commit hook.

This patch updates those instructions to use HTTPS instead of SSH for
installing the Gerrit commit hook, which should work for everyone.

Change-Id: Ibdb024aafcd601f740ae73d83c54f9659cc5a3a9
Reviewed-on: http://gerrit.cloudera.org:8080/11896
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>