De-flake sentry_client-test and sentry_authz_provider-test
Currently, sentry_client-test and sentry_authz_provider-test often fail
with 'BindException: Address already in use'. Because Sentry uses a
default port number '8038' if 'sentry.service.server.rpc-port'
is not set.

This commit sets this configuration with port 0 to ensure the mini Sentry
service can always find a free port to bound to. When running
sentry_client-test and sentry_authz_provider-test in parallel on the
same machine: before the change, one of the tests failed with
BindException; after the change, both tests passed without any errors.

Change-Id: I0e30ce88b7c2ab4afe451c50e13ef739237709e0
Reviewed-on: http://gerrit.cloudera.org:8080/11898
Tested-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>