[cmake] no empty KUDU_DATA_FILES ctest env variables
This patch introduces an update on ADD_KUDU_TEST() custom function:
if DATA_FILES is not specified for a test, don't add KUDU_DATA_FILES
variable into its ctest environment.

Prior to this patch, ctest environment of every test binary would
contain KUDU_DATA_FILES variable.  If DATA_FILES was specified for a
test, it was set to corresponding value, otherwise it was set to
an empty string.

Change-Id: I2b255f6148649cd282bab70fb19c3321fec01adb
Reviewed-on: http://gerrit.cloudera.org:8080/11907
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>