gradle: expose testRandomSeed property
In order to read a system property in tests, it isn't enough to add a
System.getProperty() call to the test; we need to expose the property in
gradle so that it's properly passed through to the test. With this patch,
testRandomSeed can be set using either -D or -P.

Note that even with testRandomSeed overridden, the PRNG didn't produce
deterministic results in TestKuduBackup.testRandomBackupAndRestore, but
perhaps that's a quirk of Scala's Random class.

Change-Id: I4873d27998f770a45dd9cc85f84ec8c146261b3d
Reviewed-on: http://gerrit.cloudera.org:8080/11908
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Adar Dembo <adar@cloudera.com>