KUDU-2599: fix flaky DefaultSourceTest.testSocketReadTimeoutPropagation
This test is just about checking the propagation of an option from the
DefaultSource to the KuduRelation. However, building a DataFrame involves a
connection to the cluster, and if we set socketReadTimeoutMs to 1, it's
going to be very difficult to connect, especially if we're testing
TSAN-instrumented binaries whose thread creation is inherently slower.

The fix? Use a magic number larger than '1'. It's that simple.

Change-Id: I81a20f776c115caf3bd720bc532bbeb5fc69bc42
Reviewed-on: http://gerrit.cloudera.org:8080/11916
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>