[location_awareness] Assign locations to clients
This patch makes it so the client is assigned a location by each master
as part of the ConnectToMaster RPC (which is part of the
ConnectToCluster process). The client will store the location it is
assigned by the leader master and potentially update it every time it
reconnects to the cluster.

This assignment will be used in a follow-up to implement location-aware
semantics for CLOSEST_REPLICA scans.

Change-Id: I0efb327293d86168a30b05305f69d011ad15587a
Reviewed-on: http://gerrit.cloudera.org:8080/11923
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>