test: deflake TestRestartWithPendingCommitFromFailedOp
The test would fail to create the tablet due to the injected failure,
when it should have really waited until the tablet was bootstrapped
before injecting errors.

The test failed 6/1000 times with stress and with a seed of -922249903.
With the fix, 1000/1000 passed.

This isn't the only source of flakiness in ts_recovery-itest, but it's
something.

Change-Id: Ic76b5f120aee7327f6cee7324b862bb89c5662ab
Reviewed-on: http://gerrit.cloudera.org:8080/11929
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>