De-flake sentry tests
Currently, sentry_client-test and sentry_authz_provider-test are still
flaky due to possible port collision. To avoid such race condition, this
patch moves the util function that finds the bind address for a daemon
based on binding mode to test_util, and updates sentry-test-base to use
it to pick up a unique bind address.

Change-Id: Id112634ad5e6f0e2c48b7e0e89389faff4f5656b
Reviewed-on: http://gerrit.cloudera.org:8080/11961
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins