[tools] Don't count tombstoned replicas in `kudu remote_replica check`
Counting them as not running made the tool pretty useless on any
realistic cluster, where it's expected to have some tombstones. For
example, it wasn't possible to use this tool to tell when all the tablet
replicas on a tablet server had finished bootstrapping.

I tested this against a server that previously failed the check because
it had about 6000/10000 replicas tombstoned in STOPPED and INITIALIZED
states. It now passes the check.

Change-Id: Ibbdde768cce5f1d27064b7276944a585b41a2f03
Reviewed-on: http://gerrit.cloudera.org:8080/11962
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>