generic_iterators: short-circuit MergeIterState::PullNextBlock
This is a micro-optimization to avoid testing every bit in the selection
vector when we've already counted up the number of set bits.

There was a lack of selection vector coverage in the existing fuzzy merge
tests, so I've modified them to randomly use one, and added a new test too.

Change-Id: I529e632f40943962d34c184b29b5cff7604ad59b
Reviewed-on: http://gerrit.cloudera.org:8080/12157
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>