java: add support for flaky test reporting
This patch hooks into the existing RetryRule to report test results to
the flaky test server inline as the tests are executed. All of the
actual reporting logic is factored out into a separate ResultReporter class.

The interface for the test reporter to pass relevant information about
the build environment to the flaky test server is based on environment
variables. This includes configuration and build metadata such as flaky
test server address, git revision, build id, build host, and build type.

This patch also includes a simple integration test for the reporter
using a mocked-up flaky test server HTTP endpoint.

This patch does not integrate the above functionality into the build.
That will happen in a follow-up patch.

Change-Id: I34f88363dbf52c2f7bba50fbfb5de059f88e7f74
Reviewed-on: http://gerrit.cloudera.org:8080/12042
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>