[spark] Add a format alias of “kudu”
Implements the spark DataSourceRegister on the
kudu-spark DefaultSource to allow the use of
the alias “kudu” when using the Kudu format.

This patch also updates the documentation examples
to use the new aliased syntax given that is the Spark
defacto approach.

Additionally the old package object that was
effectively an alias for the longer format syntax
is marked as deprecated. This syntax doesn’t match
other spark examples and sources and could be
confusing.

Change-Id: I6bb2c7314652456b2f0fa3c4a110ef156503ef92
Reviewed-on: http://gerrit.cloudera.org:8080/12083
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins