KUDU-2640: Add Spark Structured Streaming Sink
This patch adds a KuduSink and implements the
Spark StreamSinkProvider interface to support
structured streaming writes to Kudu.

These writes behave the same as writes performed
via the InsertableRelation interface.

Note: The StreamSinkProvider interface is
marked as experimental and unstable, but it has existed
since Spark 2.0.0 and is used by the Kafka integration.
Additionally, per SPARK-26415, there will be no
more Spark 2.x minor releases and Spark would
not break this API in a maintenance release. This
means the interface is effectively stable for all
of Spark 2.

Change-Id: I731e35f82c8cca7d911e4d879aa6853112132b17
Reviewed-on: http://gerrit.cloudera.org:8080/12087
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>