[client] clang: explicit override in C++11 mode only
While playing with the Kudu C++ example application, I noticed that
clang emits a lot of compilation warnings about using the explicit
override for virtual methods.  This patch addresses that, updating
the OVERRIDE macro to take into account the version of the C++ standard
the compiler is using while compiling a Kudu client application.

Change-Id: I2e11c6ce830e0b525a1b92af126f10de745c986a
Reviewed-on: http://gerrit.cloudera.org:8080/12093
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>