[tests] add /threadz pages to PeriodicWebUIChecker
With this patch, the PeriodicWebUIChecker periodically fetches
the /threadz and /threadz?group=all pages of the embedded Web server
along with other diagnostic pages.  Also, shuffle the URLs that
are being check sequentially -- that adds more randomization in
the pattern of the requests sent by the checker.

In addition, this patch updates the webserver-stress-itest to
run more readers/writer threads for TestWorkload -- that allows
to expose more race conditions in TSAN/ASAN builds (and that
is used in a follow-up changelist).

Also, this patch contains other minor/non-functional updates.

Change-Id: Iad6cbfe5108ac99231b82f662cee56ffcd25f026
Reviewed-on: http://gerrit.cloudera.org:8080/12114
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>