Support location awareness in READ_CLOSEST for the C++ client
Previously, in READ_CLOSEST, the C++ client would choose to read from a
local tablet server if possible and otherwise pick a random replica.
This changes and enhances READ_CLOSEST mode to work as follows:

1. If there is a local server, use it. If there are multiple local
   servers, choose one at random.
2. If there is a server in the same location as the client, use it. If
   there are multiple servers in the same location, choose one at
   random.
3. Otherwise, choose a server at random.

This is not a breaking change, as in the absence of locations the
behavior is consistent with what was documented before, which was only
that a local server would be chosen if possible, else a random server.

Change-Id: I2c6bcc7479c5cf2e17cb6e368ca89a1eb7f21713
Reviewed-on: http://gerrit.cloudera.org:8080/12138
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>