KUDU-2652: deflake TsRecoveryITest.TestNoBlockIDReuseIfMissingBlocks
The test waits for a write workload to generate some orphaned blocks.
With the default write pattern, this could sometimes take a while and
cause the test to hit an error. This patch makes orphaned block
generation faster by making the workload update a single row.

Before this, I looped the test in debug mode with 32 stress threads and
it failed 15/100 times. With the fix, this passed 1000/1000.

Change-Id: I24a689702d29a744be8c113fedafec0307a90b1c
Reviewed-on: http://gerrit.cloudera.org:8080/12166
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>