[gutil] suppress -Wdeprecated-declarations warning on macOS
As it turned out, most of the atomic functions used in
atomicops-internals-macosx.h have been declared deprecated since
macOS 10.12, most of the functions in the SASL API are deprecated
since macOS 10.11, and the krb5 API is deprecated in favor
of GSS.framework.

To avoid multiple compilation warnings on macOS while compiling
with clang from the Xcode development environment, this changelist adds
'pragma GCC diagnostic push/pop' for the '-Wdeprecated-declarations'
flag accordingly.  It's confirmed that this patch suppresses
deprecation warnings at least on macOS 10.14.

This is a follow-up to 678dbac6fb05d0370e40e6645d4b1ec530fa0180.

Change-Id: I82b0ebd9917a567d1a8b72c80b47dc102304d860
Reviewed-on: http://gerrit.cloudera.org:8080/12183
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins