generic_iterators: prep for MergeIterator dominance
This patch adds a counter to the MergeIterator that tracks the number of
comparisons made during the lifetime of the iterator. When the dominance
algorithm is added, the counter's value ought to drop. TestMerge now logs
the counter's value, and can also generate non-overlapped inputs if desired.

The counter isn't exposed to users. It could be added to IteratorStats and
set for every key column, but even then it'll only apply to ORDERED scans,
so the value is dubious.

Change-Id: I3a8db5973dc09b7271e05b3cc28025b7a2a9e21b
Reviewed-on: http://gerrit.cloudera.org:8080/12196
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins