[java] log warning if applying operation on a closed session
Closed sessions won't flush on another close() call, so it's somewhat
dubious to apply() an operation to a closed session. This change adds a
warning if apply() is used with a closed session.

Originally this patch enforced the new behavior via precondition, but that's
a breaking change for clients that relied on the old broken behavior.

Change-Id: I0fe221fedbb91959985f5ee374f1b691be2426a9
Reviewed-on: http://gerrit.cloudera.org:8080/12237
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>