switch all iterators to unique_ptr
We've suffered from a longstanding wart in the iterator subsystem: a mix of
raw and various kinds of smart pointers. The usage of shared_ptr was
particularly vexing as it suggested that iterators had shared ownership when
in fact they didn't.

This yak shave of a patch addresses all of those issues by converting all of
the iterator pointer types to unique_ptr. I snuck in some clang-tidy
suggestions too, but otherwise the cleanup here should narrowly scoped.

Change-Id: I60cdc49e6a209c72cb33681ae1be89586d739204
Reviewed-on: http://gerrit.cloudera.org:8080/12222
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>