[docs] Add docs for rack/location-awareness
I stuck to the name "rack-awareness" throughout, preferring it to the
term we have used internally, "location-awareness", because even though
the latter term is more technically correct, the former is the more
typical and colloquial name for this sort of thing.

These docs borrow heavily from the WIP blog post written by Alexey.

Change-Id: I505b7fd545a4253170a5b7cac63c33e7d9669490
Reviewed-on: http://gerrit.cloudera.org:8080/12219
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>