KUDU-2662 TestKuduClient.testClientLocation is flaky
The flakiness is caused by a difference in how Python 2.6 requires the
placeholders to string.format to be sepcified. See
https://stackoverflow.com/a/21035037.

There's likely a second source of flakiness because the code that is
breaking is about to throw an exception! So, expect a follow up
addressing the true source of flakiness. To help diagnose, a bit of
extra information is added to the exception the flaky tests were trying
to throw.

Change-Id: I0b37d799732307d839853758681b1be820f6e0dd
Reviewed-on: http://gerrit.cloudera.org:8080/12236
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>