[fs]: delete the (orphaned) metadata file when repairing
With KUDU-2636 fixed, we now support deleting dead containers at
runtime. Because this involves deleting a pair of files, there's
a time window in which it's possible for there to be just one file.
A well-timed crash or kill -9 can make this a permanent state that
fails the log block manager at startup. Indeed, a loop of
dense_node-itest failed 2/1000 times in this way.

This patch fixes the problem by detecting "orphaned" metadata files
at startup and deleting them. A metadata file is orphaned if it has
no corresponding data file and if it contains nothing but dead blocks.

Change-Id: I2a4223d0e6385a7f7776f8fe2679a3bb16d832ab
Reviewed-on: http://gerrit.cloudera.org:8080/12239
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins