[spark] Add metrics to kudu-spark
This adds some basic accumulator metrics to kudu-spark. These
accumulators appear for each stage involving a KuduContext, with values
broken down by executor.

Follow-ups will add some more sophisticated metrics and use metrics to
add additional logging.

In addition to the unit tests, I tested this on a 3-node cluster and
confirmed I was able to see the accumulator values for stages and
individual tasks on the Spark application's web UI, and that were
correct.

Change-Id: Ied81c890b3d4510f767d8f023963ff878f398140
Reviewed-on: http://gerrit.cloudera.org:8080/12251
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Will Berkeley <wdberkeley@gmail.com>